+++ src/lib/language-support.ts
@@ -14,6 +14,7 @@ const languages: Map<string, LanguageDefinition> = new Map(
         defineLanguage("c"),
         defineLanguage("clojure"),
         defineLanguage("css"),
+        defineLanguage("fennel"),
         defineLanguage("html"),
         defineLanguage("java"),
         defineLanguage("javascript"),